Add files via upload
I spent time collecting data from data.gov, cleaned and reformatted the data, and used PivotTable and PivotCharts to design the dashboard. The dashboard is an interactive dashboard that can be selected by readers. I also attached the data dictionary and source in the last two tabs. The dashboard can provide the correlations of Age, Education, Income, race-ethnicity, and gender vs. Depression. Hope you enjoy using the dashboard as a template or using it for other interactive Excel dashboard projects.